% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_ripser}
\alias{run_ripser}
\title{Ripser}
\usage{
run_ripser(dist_mat, dim = 1L, threshold = NA_real_)
}
\arguments{
\item{dist_mat}{distance matrix}

\item{dim}{compute persistent homology up to this dimension}

\item{threshold}{compute Rips complex up to this diameter}
}
\description{
Run ripser
}
\examples{
rcircle <- function(N, r = 1, sd = 0){
radius <- rnorm(N, r, sd)
angle <- runif(N, 0, 2*pi)
data.frame(x = radius * cos(angle), 
           y = radius * sin(angle))
}
dta <- rcircle(200, sd = 0.1)
dmat <- as.matrix(dist(dta))
pers <- run_ripser(dmat, dim = 2)
}
